Fixed errors on number og blocks being stored
Fixed function declaration that had the wrong return size. This had the effect
of discarding significant bytes from the number of blocks in an sqz file.

Block counter moved to outside sqzlib API

+++ src/sqzlib/sqzlib.h
@@ -170,7 +170,7 @@ uint64_t sqz_filesize(sqzFile sqzfp);
 /*
   Get number of blocks in sqz file
 */
-uint8_t sqz_getblocks(sqzFile sqzfp);
+uint64_t sqz_getblocks(sqzFile sqzfp);
 
 
 /*
